separated model from view
