+++ Cargo.toml
@@ -12,5 +12,4 @@ piston2d-graphics = "0.45.0"
 piston_window = "0.132.0"
 image = "0.25.2"
 rand = "0.9.1"
-gfx_device_gl = "0.16.2"
 find_folder = "0.3.0"
